net: usb: ipheth: add CDC NCM support (PoC)
Recent iOS releases support CDC NCM encapsulation on RX. This mode is
the default on macOS and Windows.

When reconnecting a device from a macOS/Windows machine to a Linux
machine, the device stays in NCM mode. Therefore to correctly support
such a device, the driver has to either support the NCM mode too, or
somehow put the device back into legacy mode.

To match the behaviour of the macOS/Windows driver, and since there
are no documented control commands to revert to legacy mode, I chose
to implement basic NCM support.

With this change, the device is attempted to be put into NCM mode by
default, and falls back to legacy mode if said attempt failed.

Signed-off-by: Foster Snowhill <Forst@users.noreply.github.com>
Signed-off-by: Georgi Valkov <gvalkov@gmail.com>