coreutils: Disable sparse copy on macOS
Due to a bug in macOS, sparse copies are corrupted on virtual disks
formatted with APFS. HFS is not affected. Affected are coreutils
install, and gcp when compiled with SEEK_HOLE, as well as macOS Finder.

While reading the entire file returns valid data, scanning for
allocated segments may return holes where valid data is present.
In this case a sparse copy does not contain these segments and return
zeroes instead. Once the virtual disk is dismounted and then
mounted again, a sparse copy produces correct results.

This breaks OpenWRT build on macOS. Steps to reproduce:
https://httpstorm.com/share/.openwrt/test/2023-02-06_coreutils-9.1/

Signed-off-by: Georgi Valkov <gvalkov@gmail.com>