tools/coreutils: update to 9.5
The coreutils archive is already bootstraped and ready to build
after ./configure. Bootstraping with the wrong version of gnulib
results in build errors on macOS and is not necessary, unless we
need to patch the ./bootstrap script.

./lib/obstack.h:143:5: error: invalid token at start of a preprocessor expression
 #if @SMALL_PTRDIFF_T@

Reverts part of 09f8f21ae33402201c14a125a37b64746415642d
tools/coreutils: bootstrap to local gnulib source

Signed-off-by: Georgi Valkov <gvalkov@gmail.com>