usbnet: ipheth: prevent OoB reads of NDP16 and DPEs
There were a few potential out of bounds reads when processing
malformed URBs received from a connected device in NCM mode:

* Only the start of NDP16 (wNdpIndex) was checked to fit in the URB
  buffer.
* Datagram length check as part of DPEs could overflow.
* DPEs could be read past the end of NDP16 and even end of URB buffer
  if a trailer DPE wasn't encountered.

The above is not expected to happen in normal device operation.

To address the above issues without reimplementing more of CDC NCM,
rely on and check for a specific fixed format of incoming URBs:

* 12-byte NTH16
* 96-byte NDP16, allowing up to 22 DPEs (up to 21 datagrams + trailer)

The length of NDP16 is constant regardless of the DPE count. NDP16
directly follows NTH16.

Adapt the driver to use the fixed URB format. Allow up to 22 DPEs.
Always expect a null trailer DPE. The minimal URB length of 108 bytes
(IPHETH_NCM_HEADER_SIZE) in NCM mode is enforced since introduction of
NCM support.

Signed-off-by: Foster Snowhill <forst@pen.gy>
Signed-off-by: Georgi Valkov <gvalkov@gmail.com>